HHH-13726 Extract 'prepareEventListeners' method out of SessionFactoryImpl
