+++ hibernate-core/src/main/java/org/hibernate/boot/spi/MetadataImplementor.java
@@ -6,13 +6,12 @@
  */
 package org.hibernate.boot.spi;
 
-import java.util.Collection;
 import java.util.Set;
 
 import org.hibernate.MappingException;
 import org.hibernate.boot.Metadata;
-import org.hibernate.cache.cfg.internal.DomainDataRegionConfigImpl;
 import org.hibernate.engine.spi.Mapping;
+import org.hibernate.engine.spi.SessionFactoryImplementor;
 import org.hibernate.internal.SessionFactoryImpl;
 import org.hibernate.mapping.MappedSuperclass;
 import org.hibernate.query.spi.NamedQueryRepository;
@@ -59,4 +58,6 @@ public interface MetadataImplementor extends Metadata, Mapping {
 	void validate() throws MappingException;
 
 	Set<MappedSuperclass> getMappedSuperclassMappingsCopy();
+
+	void initSessionFactory(SessionFactoryImplementor sessionFactoryImplementor);
 }
